+++ control-plane/mbus-api/src/lib.rs
@@ -294,6 +294,10 @@ pub enum ResourceKind {
     Pool,
     /// Replica resource
     Replica,
+    /// Replica state
+    ReplicaState,
+    /// Replica spec
+    ReplicaSpec,
     /// Nexus resource
     Nexus,
     /// Child resource
