+++ control-plane/store/src/types/v0/volume.rs
@@ -1,0 +1,153 @@
+//! Definition of volume types that can be saved to the persistent store.
+
+use crate::{
+    store::{ObjectKey, StorableObject, StorableObjectType},
+    types::SpecState,
+};
+use mbus_api::{v0, v0::VolumeId};
+use serde::{Deserialize, Serialize};
+
+type VolumeLabel = String;
+
+/// Volume information
+#[derive(Serialize, Deserialize, Debug, PartialEq)]
+pub struct Volume {
+    /// Current state of the volume.
+    pub state: Option<VolumeState>,
+    /// Desired volume specification.
+    pub spec: VolumeSpec,
+}
+
+/// Runtime state of the volume.
+/// This should eventually satisfy the VolumeSpec.
+#[derive(Serialize, Deserialize, Debug, PartialEq)]
+pub struct VolumeState {
+    /// Volume Id
+    pub uuid: v0::VolumeId,
+    /// Volume size.
+    pub size: u64,
+    /// Volume labels.
+    pub labels: Vec<VolumeLabel>,
+    /// Number of replicas.
+    pub num_replicas: u8,
+    /// Protocol that the volume is shared over.
+    pub protocol: v0::Protocol,
+    /// Nexuses that make up the volume.
+    pub nexuses: Vec<v0::NexusId>,
+    /// Number of front-end paths.
+    pub num_paths: u8,
+    /// State of the volume.
+    pub state: v0::VolumeState,
+}
+
+/// Key used by the store to uniquely identify a VolumeState structure.
+pub struct VolumeStateKey(VolumeId);
+
+impl From<&VolumeId> for VolumeStateKey {
+    fn from(id: &VolumeId) -> Self {
+        Self(id.clone())
+    }
+}
+
+impl ObjectKey for VolumeStateKey {
+    fn key_type(&self) -> StorableObjectType {
+        StorableObjectType::VolumeState
+    }
+
+    fn key_uuid(&self) -> String {
+        self.0.to_string()
+    }
+}
+
+impl StorableObject for VolumeState {
+    type Key = VolumeStateKey;
+
+    fn key(&self) -> Self::Key {
+        VolumeStateKey(self.uuid.clone())
+    }
+}
+
+/// User specification of a volume.
+#[derive(Serialize, Deserialize, Debug, Clone, PartialEq)]
+pub struct VolumeSpec {
+    /// Volume Id
+    pub uuid: v0::VolumeId,
+    /// Size that the volume should be.
+    pub size: u64,
+    /// Volume labels.
+    pub labels: Vec<VolumeLabel>,
+    /// Number of children the volume should have.
+    pub num_replicas: u8,
+    /// Protocol that the volume should be shared over.
+    pub protocol: v0::Protocol,
+    /// Number of front-end paths.
+    pub num_paths: u8,
+    /// State that the volume should eventually achieve.
+    pub state: VolumeSpecState,
+    /// Update of the state in progress
+    #[serde(skip)]
+    pub updating: bool,
+}
+
+/// Key used by the store to uniquely identify a VolumeSpec structure.
+pub struct VolumeSpecKey(VolumeId);
+
+impl From<&VolumeId> for VolumeSpecKey {
+    fn from(id: &VolumeId) -> Self {
+        Self(id.clone())
+    }
+}
+
+impl ObjectKey for VolumeSpecKey {
+    fn key_type(&self) -> StorableObjectType {
+        StorableObjectType::VolumeSpec
+    }
+
+    fn key_uuid(&self) -> String {
+        self.0.to_string()
+    }
+}
+
+impl StorableObject for VolumeSpec {
+    type Key = VolumeSpecKey;
+
+    fn key(&self) -> Self::Key {
+        VolumeSpecKey(self.uuid.clone())
+    }
+}
+
+/// State of the Volume Spec
+pub type VolumeSpecState = SpecState<v0::VolumeState>;
+
+impl From<&v0::CreateVolume> for VolumeSpec {
+    fn from(request: &v0::CreateVolume) -> Self {
+        Self {
+            uuid: request.uuid.clone(),
+            size: request.size,
+            labels: vec![],
+            num_replicas: request.replicas as u8,
+            protocol: v0::Protocol::Off,
+            num_paths: request.nexuses as u8,
+            state: VolumeSpecState::Creating,
+            updating: true,
+        }
+    }
+}
+impl PartialEq<v0::CreateVolume> for VolumeSpec {
+    fn eq(&self, other: &v0::CreateVolume) -> bool {
+        let mut other = VolumeSpec::from(other);
+        other.state = self.state.clone();
+        other.updating = self.updating;
+        &other == self
+    }
+}
+impl From<&VolumeSpec> for v0::Volume {
+    fn from(spec: &VolumeSpec) -> Self {
+        Self {
+            uuid: spec.uuid.clone(),
+            size: spec.size,
+            state: v0::VolumeState::Unknown,
+            children: vec![],
+        }
+    }
+}
